Moved utilities to their own folder
